Remove getValue default & valueFromMaybe
Prior to this commit getValue always returned a string, even when the
field was not present, using an empty string in that case.  This
prevents the caller from being able to determine whether the field was
not present, which is sometimes required.  Return the Maybe instead.

Since getValue was the only user of valueFromMaybe, remove that as well.
Callers can use fromMaybe from Data.Maybe to supply a default value if
they need that functionality.  This is how the existing tests were fixed
up to work without changing them significantly.