Improve key-value stripping in msgfmt
The msgfmt module previously stripped whitespace from both ends of keys
and values when inserting a pair with insertKV or when parsing a form;
this was likely set up due to the fact that PacForms does this as well.
It appears that the main reason for this behavior is that form keys
often (but not always) have an extraneous newline at the beginning due
to the textual representation; stripping whitespace from the keys takes
care of this.  However, this behavior can also cause information to be
lost when round-tripping messages through the textual representation ---
or even when simply converting a message from some other format to the
Msg datatype and back --- so this commit changes the msgfmt parsing code
to only strip leading newlines when necessary.