Make ICS213.Msg Body polymorphic
This commit improves the ICS213 handling of body fields in an ICS213
form: instead of unconditionally parsing the body fields to/from a map,
and requiring form-specific functions to build on top of this map, the
ICS213 module now provides a typeclass, ICS213Body, that allows other
modules to provide custom implementations of the body part of an ICS 213
form: the body field of an ICS213.Msg can be of any type implementing
this typeclass.  This new functionality should, in the future, make it
easier to implement functions specific to other form types.

Error handling for message bodies is not yet implemented, but is planned
for a future commit.