Extract parsing utilities into ParseUtils module
The ICS213 parsing routines previously included a number of utility
functions for parsing in where clauses for parsing functions; the
FormatError datatype was also declared in the ICS213 module.  As these
datatypes and functions can be useful when implementing message body
parsing, this commit creates a new PackItForms.ParseUtils module and
extracts these utilities into it.