Add support for "rolling up" BAT statuses
This commit adds two functions to PackItForms.LADamage:
rollupBATStatuses and addBATStatuses.  The former is used for rolling up
a list of BAT status reports into a total, up-to-date status report; the
latter is used for adding two BAT status reports, and is used in the
implementation of the former.  addBATStatuses is intended to be used
when attempting to keep an aggregate status record up-to-date as new
information appears.