Add phantom type parameter to LADamage BATStatus
The BATStatus field which contains most of the data conveyed by a Los
Altos Damage Assessment form has two possible semantic meanings: it can
represent an incremental update on the status of a BAT, or it can
contain the total, aggregate, best-at-this-time information representing
everything that is known about a BAT.  To clarify this semantic
difference, this commit adds a phantom type parameter to the BATStatus
record that can be used to indicate which purpose a value is being used
for.  The parameter is intended to be one of Update or Accumulated.