Fix role handling for ICS213
Previously, the ICS213 library did not correctly embody the semantics
behind the different message numbers sent with each form, in part
because MsgFmt did not correctly expose functions for handling the
Outpost message envelope.  This commit modifies ICS213 to correctly
determine the role of a message based on the various message number
fields that may be included with it.