Add functions for manipulating zone mappings
Mappings from zones to the contained BATs and from a BAT to the
containing zone are needed in a bunch of places.  Prior to this commit
there was a type for a data structure to maintain that information but
the user had to pattern match with the value constructor and use
functions on the contained data types which was inconvenient and
verbose.  This commit adds convenience functions to clean that up.