Remove descriptive field name trailers in LADamage
For human comprehension, the fields in pack-it-forms forms allow an
optional trailer on numbered fields consisting of a period followed by
text.  The damage assessment code expected these descriptive field name
trailers in the keys of messages being parsed and output them in
generated messages.  However, that is actually incorrect as the
pack-it-forms does not transmit them on the radio to minimize message
size.  This commit removes them.