More sophisticated nix build infrastructure
This makes several changes to the nix build infrastructure, which
should make it less brittle and less work to maintain.  Nix now
automatically generates the Nix recipe for the package from the Cabal
file using callCabal2Nix, and also properly filters the source
directory so that nix-build result symlinks do not force package
rebuilds.  The nix infrastructure now supports building in both
"release" (the default when using release.nix) and "dev" (the default
when using default.nix or shell.nix) modes; currently, "release" mode
results in fetching non-Hackage dependencies from GitHub, while "dev"
expects them to be local.  The default GitHub and local paths for
dependencies are specified in project-deps.nix, and can be overriden
on the commandline:

nix build -f release.nix --arg args "{ dep1 = /path/to/dep1; ... }"

Release mode also builds with a pinned nixpkgs, while dev mode builds
with the system's local nixpkgs, at least for now, to encourage
keeping up-to-date with upstream nixpkgs changes.

To add a new Hackage dependency, it must be added to the Cabal file.
To override a Hackage dependecy version, add a line like this to
project-deps.nix:

  dependency = self.hsp.callHackage "dependency-name" "version" {};

To add a new local dependency which uses the same structure as this
project, and which is fetched from GitHub in release mode and imported
from a local copy in dev mode, add a section like this to
project-deps.nix:

  dependency = self.releaseGHDevLocal {
    owner = "GH owner";
    repo = "GH repo";
    rev = "Release rev";
    sha256 = "Release hash";
    local = /path/to/local/dependency;
  };