Pin nixpkgs and sdtpl for release builds
nixpkgs is pinned to the tip of the 17.09 branch at the time of this
commit. The intention is to keep it on the most recent stable commit
at each release.

sdtpl is pinned to the tip of master at the time of this commit. It is
not expected to change significantly unless changes in dependencies
will require updates.